pyproject.toml
[project]
name = "gaas"
version = "0.1.0"
description = "AI-powered personal assistant — email, PRs, issues, safely"
requires-python = ">=3.11"
dependencies = [
    "gaas-sdk",
    "gaas-email",
    "gaas-github",
    "fastapi[standard]>=0.129.0",
    "jinja2>=3.1.6",
    "jsonschema>=4.23",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0",
    "fastapi-crons>=0.2",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
gemini = ["gaas-gemini"]

[project.scripts]
gaas = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "hypothesis>=6.100",
    "mypy>=1.10",
    "complexipy>=5.2.0",
    "radon>=6.0",
    "vulture>=2.11",
    "deptry>=0.16",
    "import-linter>=2.0",
    "bandit>=1.7",
    "ruff>=0.4",
    "claude-agent-sdk>=0.1.45",
]

[tool.uv.sources]
gaas-sdk = { path = "packages/gaas-sdk", editable = true }
gaas-email = { path = "packages/gaas-email", editable = true }
gaas-github = { path = "packages/gaas-github", editable = true }
gaas-gemini = { path = "packages/gaas-gemini", editable = true }

[tool.pytest.ini_options]
testpaths = ["tests", "packages/gaas-email/tests", "packages/gaas-gemini/tests", "packages/gaas-github/tests", "packages/gaas-sdk/tests"]
pythonpath = ["."]
addopts = "--import-mode=importlib"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
exclude = ["packages/gaas-gemini/"]
explicit_package_bases = true
mypy_path = ["packages/gaas-sdk/src", "packages/gaas-email/src"]

[[tool.mypy.overrides]]
module = "gaas_sdk.*"
strict = true

[[tool.mypy.overrides]]
module = "app.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = [
    "imap_tools.*",
    "icalendar.*",
    "frontmatter.*",
    "fastapi_crons.*",
    "ruamel.*",
    "complexipy.*",
    "bs4.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "ANN", "UP", "RUF", "B", "SIM"]
ignore = ["ANN401", "ANN204"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN"]
"packages/*/tests/**/*.py" = ["ANN"]
"app/**/*.py" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202"]
"packages/gaas-email/**/*.py" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202"]
"packages/gaas-github/**/*.py" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202"]
"packages/gaas-gemini/**/*.py" = ["ANN001", "ANN002", "ANN003", "ANN201", "ANN202"]

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]
